First new version of mirror calibration in firestation.
The mirror calibration expects marks on the floor in different angles
(wrt the robot) in different images. For example, the input of the
mirror calibration can be one image with the marks at 0 degrees, one
with the marks at 120 degrees and one image with the marks at 240 degrees.
The mirror calibration tries to recognize those marks and knows the real
distances of them. Then interpolation is done.

This is essentially the commit 32be547fcb01e2c5980148d68d056ca2d17d30e5
from the old repository adapted to the firevision-refactoring.