Angles of marks can be set/adjusted by user.
The GUI has a scale-component to set the position of the marks between
-180 and 180 degrees for all directions.

Fixed: bulb generation had assertion failure if the two nearest neighbors
of the considered points have both positive or both negative angles with
respect to the considered point. This scenario never occured with fix
mark angles 0, 120, 240 degrees, because then always one of the nearest
neighbors had a positive difference angle and the other had a negative
one.

This is an adapted version of 3e6758f8a294efc23497734154dc4105f6b16cb9
from the old repository.