Reduced expected marks from 9 to 7, simplified hole-filtering.
The hole of the fourth and last black sheet is difficult to recognize because
it becomes so small that the distinguishing edges after edge detection are much
thicker than the hole itself.
Hence it's probably the best just to ignore this paper.

An alternative was to put two or three black sheets in a row or so and adapt
the calculate_real_distance() appropriately.

Also removed some methods that were nowhere used.
They were used to scale holes (to compensate the effect that the further away a
hole is the smaller it looks). This was done with so many stupid magic numbers
that the methods probably were stupid.